implement Game.Map
We moved interfaces inside of Game to better illustrate where they are used.
Changed GameMap to be an interface and not a class.
The important change is making findRoute an extension function.
We do this to be able to wrap it in an Either, to allow returning both types of result.
We also create a function to easily create the options object.