add Options implementation
Adds an Options interface to let us restrict what can be created with the options function.
Interfaces that implement Options should have only var properties, and they should be nullable.
We need the Options interface to ensure typesafety.
The previous jsObject is too unsafe.
Adds a function, options, to allow for easy creation of option objects.
Kotlin automatically infers what option type is needed if the function is used as an argument to a method that takes Options.