Fix trait usage in API (#15)
* removes the Container and SpawnEnergyProvider traits

They both cause issues when trying to get potential targets from a list of structures.
It is impossible to get a list that contains both Structure and Container,
thus you could not easily use the target in multiple ways.

* make StoreDefinition a record

Allows us to easily access all Resources on properties of type StoreDefinition.
We leave `energy` as the only property, as that is definitely existing.
Other resources can then be accessed using the Record's get function.
E.g. `creep.store[RESOURCE_OXYGEN]`.