Structure interface (#40)
add structure interface as the supertype of energycontainer and Store

replace type Structure with Istructure in all api parameters because IStructure is more flexible than Structure. We keep Structure as *return type* because it is richer in terms of
operations that can be performed on it (equals, hashcode)

E.g. This allows to have val energyProviders : Array<EnergyContainer which contains both StructureSpawn and StructureExtension in a way that still allows creeps to transfer to them. The last bit was previously not possible because we could not transfer to an EnergyContainer.