add an overload for treating null as 0 in numeric comparisons
This is useful because it allows us to write
`creep.store[RESOURCE_ENERGY] < creep.store.getCapacity()`
which is possible in Javascript (but not plain Kotlin) because
in Javascript `null` is treated as `0` in numeric comparisons

Kotlin does not define such a function, likely because it does not want to assert
how null should relate to other numbers (is it smaller than anything? larger? eq 0?)
in pursuit of good interop with Screeps this library takes a stance